irshados-backend/requirements.txt
django>=5,<6
djangorestframework>=3.15
drf-spectacular>=0.27
django-cors-headers>=4.4
python-dotenv>=1.0
psycopg[binary]>=3.2
djangorestframework-simplejwt[crypto]>=5.3
celery>=5.4
redis>=5.0
django-celery-results>=2.5
